refactor(lib::config): move yaml-theme structure to own file
because that is independent of the normal config
